rl_env_demo/pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-smart-home-energy"
version = "1.0.0"
description = "Smart Home Energy Management environment for OpenEnv - optimize energy consumption with HVAC, battery, and solar"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    # install from github
    # "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git",
    "openenv-core[core]>=0.2.2",
    # Environment-specific dependencies
    "numpy>=1.20.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m rl_env_demo.server.app
server = "rl_env_demo.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["rl_env_demo", "rl_env_demo.server"]
package-dir = { "rl_env_demo" = ".", "rl_env_demo.server" = "server" }